requirements.txt
python-telegram-bot==20.7
aiosqlite==0.19.0
python-dotenv==1.0.1
